% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{INFCdomain}
\alias{INFCdomain}
\title{The range of applicability of \code{INFCvpe()} function}
\format{A data frame with 18228 observations and 2 columns}
\description{
A \code{data.frame} containing the "range of applicability" (or "domain")
of \code{INFCvpe()} function
}
\references{
Tabacchi G., Di Cosmo L., Gasparini P., Morelli S., 2011a.
\emph{Stima del volume e della fitomassa delle principali specie forestali
italiane. Equazioni di previsione, tavole del volume e tavole della
fitomassa arborea epigea. Stima del volume e della fitomassa delle
principali specie forestali italiane. Equazioni di previsione, tavole del
volume e tavole della fitomassa arborea epigea}. 412 pp. [ITA, ita]
}
\keyword{datasets}

