% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForIT.R
\docType{package}
\name{ForIT-package}
\alias{ForIT}
\alias{ForIT-package}
\title{ForIT: A package that implements as R functions the INFC-2005 equations estimating single tree volume and phytomass .}
\description{
The ForIT package provides two main functions, estimating respectively individual and summary
values, and some accessory functions, facilitating package use and documenting its content
}
\references{
Tabacchi G., Di Cosmo L., Gasparini P., Morelli S., 2011a. \emph{Stima
del volume e della fitomassa delle principali specie forestali italiane.
Equazioni di previsione, tavole del volume e tavole della fitomassa arborea
epigea. Stima del volume e della fitomassa delle principali specie
forestali italiane. Equazioni di previsione, tavole del volume e tavole
della fitomassa arborea epigea}. 412 pp.

Tabacchi G., Di Cosmo L., Gasparini P., 2011b. \emph{Aboveground tree
volume and phytomass prediction equations for forest species in Italy}.
European Journal of Forest Research 130: 6 911-934
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/NuoroForestrySchool/ForIT.git}
  \item \url{https://zenodo.org/records/5790157}
}

}
\author{
\strong{Maintainer}: Nicola Puletti \email{nicola.puletti@gmail.com} (\href{https://orcid.org/0000-0002-2142-959X}{ORCID})

Authors:
\itemize{
  \item Mirko Grotti \email{mirko.grotti@gmail.com}
  \item Roberto Scotti \email{scotti@uniss.it} (\href{https://orcid.org/0000-0001-7394-4473}{ORCID})
}

}
\keyword{internal}
