\name{INFCdomain}
\alias{INFCdomain}
\docType{data}
\title{The range of applicability of \code{INFCvpe()} function}
\description{
A \code{data.frame} containing the "range of applicability" (or "domain")
of \code{INFCvpe()} function
}

\format{
  A data frame with 18563 observations on the following 2 variables.
  \describe{
    \item{\code{key}}{a character vector}
    \item{\code{in.range}}{a factor with levels \code{} \code{y}}
  }
}
\references{
Tabacchi G., Di Cosmo L., Gasparini P., Morelli S., 2011a.
\emph{Stima del volume e della fitomassa delle principali specie forestali
italiane. Equazioni di previsione, tavole del volume e tavole della
fitomassa arborea epigea. Stima del volume e della fitomassa delle
principali specie forestali italiane. Equazioni di previsione, tavole del
volume e tavole della fitomassa arborea epigea}. 412 pp. [ITA, ita]
}
\keyword{datasets}
